split up Azure connected tests
